Moved main file back to main.py instead of sprite-test.py
and I added a key to close the program
