--- testing.py
@@ -1,282 +1,0 @@
-
-
-
-import sys
-import random
-import math
-import os
-import getopt
-import pygame
-from socket import *
-from pygame.locals import *
-
-
-
-class Player(pygame.sprite.Sprite):
-    def __init__(self):
-        pygame.sprite.Sprite.__init__(self)
-        self.image = pygame.image.load("intro_ball.gif")
-        self.rect = self.image.get_rect()
-        screen = pygame.display.get_surface()
-        self.area = screen.get_rect()
-        self.speed = 10
-        self.state = "still"
-        self.reinit()
-
-    def reinit(self):
-        self.state = "still"
-        self.movepos = [0,0]
-        
-        self.rect.midleft = self.area.midleft
-
-    def update(self):
-        newpos = self.rect.move(self.movepos)
-        if self.area.contains(newpos):
-            self.rect = newpos
-        pygame.event.pump()
-
-    def moveup(self):
-        self.movepos[1] = self.movepos[1] - (self.speed)
-        self.state = "moveup"
-
-    def movedown(self):
-        self.movepos[1] = self.movepos[1] + (self.speed)
-        self.state = "movedown"
-
-    def moveright(self):
-        self.movepos[0] = self.movepos[0] + (self.speed)
-        self.state = "moveright"
-
-    def moveleft(self):
-        self.movepos[0] = self.movepos[0] - (self.speed)
-        self.state = "moveleft"
-
-    def stopmovevert(self):
-        self.movepos[1] = 0
-        self.state = "still"
-
-    def stopmovehoriz(self):
-        self.movepos[0] = 0
-        self.state = "still"
-    
-
-
-def main():
-# Initialise screen
-    pygame.init()
-    screen = pygame.display.set_mode((640, 480))
-    pygame.display.set_caption('Basic Pong')
-
-    # Fill background
-    background = pygame.Surface(screen.get_size())
-    background = background.convert()
-    background.fill((0, 0, 0))
-
-    # Initialise players
-    global player1
-    player1 = Player()
-    
-
-    # Initialise ball
-   
-
-    # Initialise sprites
-    playersprites = pygame.sprite.RenderPlain((player1))
-    
-
-    # Blit everything to the screen
-    screen.blit(background, (0, 0))
-    pygame.display.flip()
-
-    # Initialise clock
-    clock = pygame.time.Clock()
-
-    # Event loop
-    while 1:
-        # Make sure game doesn't run at more than 60 frames per second
-        clock.tick(60)
-
-        for event in pygame.event.get():
-            if event.type == QUIT:
-                return
-            elif event.type == KEYDOWN:
-                if event.key == K_w:
-                    player1.moveup()
-                if event.key == K_s:
-                    player1.movedown()
-                if event.key == K_a:
-                    player1.moveleft()
-                if event.key == K_d:
-                    player1.moveright()
-                
-            elif event.type == KEYUP and (not event.type == KEYDOWN):
-                #if event.key == K_a or event.key == K_w or event.key == K_s or event.key == K_d:
-                #    player1.movepos = [0,0]
-                #    player1.state = "still"
-
-                if event.key == K_a or event.key == K_d:
-                    player1.stopmovehoriz()
-
-                if event.key == K_w or event.key == K_s:
-                    player1.stopmovevert()
-                    
-                
-
-        
-        screen.blit(background, player1.rect, player1.rect)
-        
-        playersprites.update()
-        playersprites.draw(screen)
-        pygame.display.flip()
-
-
-if __name__ == '__main__': main()
-
-while 1:
-    for event in pygame.event.get():
-        if event.type == pygame.QUIT: sys.exit()
-
-    #screen.fill(black)
-    #screen.blit(ball, ballrect)
-
-
-
-import sys
-import random
-import math
-import os
-import getopt
-import pygame
-from socket import *
-from pygame.locals import *
-
-
-
-class Player(pygame.sprite.Sprite):
-    def __init__(self):
-        pygame.sprite.Sprite.__init__(self)
-        self.image = pygame.image.load("intro_ball.gif")
-        self.rect = self.image.get_rect()
-        screen = pygame.display.get_surface()
-        self.area = screen.get_rect()
-        self.speed = 10
-        self.state = "still"
-        self.reinit()
-
-    def reinit(self):
-        self.state = "still"
-        self.movepos = [0,0]
-        
-        self.rect.midleft = self.area.midleft
-
-    def update(self):
-        newpos = self.rect.move(self.movepos)
-        if self.area.contains(newpos):
-            self.rect = newpos
-        pygame.event.pump()
-
-    def moveup(self):
-        self.movepos[1] = -1
-        self.state = "moveup"
-
-    def movedown(self):
-        self.movepos[1] = 1
-        self.state = "movedown"
-
-    def moveright(self):
-        self.movepos[0] = 1
-        self.state = "moveright"
-
-    def moveleft(self):
-        self.movepos[0] = -1
-        self.state = "moveleft"
-
-    def stopmovevert(self):
-        self.movepos[1] = 0
-        self.state = "still"
-
-    def stopmovehoriz(self):
-        self.movepos[0] = 0
-        self.state = "still"
-    
-
-
-def main():
-# Initialise screen
-    pygame.init()
-    screen = pygame.display.set_mode((640, 480))
-    pygame.display.set_caption('Basic Pong')
-
-    # Fill background
-    background = pygame.Surface(screen.get_size())
-    background = background.convert()
-    background.fill((0, 0, 0))
-
-    # Initialise players
-    global player1
-    player1 = Player()
-    
-
-    # Initialise ball
-   
-
-    # Initialise sprites
-    playersprites = pygame.sprite.RenderPlain((player1))
-    
-
-    # Blit everything to the screen
-    screen.blit(background, (0, 0))
-    pygame.display.flip()
-
-    # Initialise clock
-    clock = pygame.time.Clock()
-
-    # Event loop
-    while 1:
-        # Make sure game doesn't run at more than 60 frames per second
-        clock.tick(60)
-
-        for event in pygame.event.get():
-            if event.type == QUIT:
-                return
-            elif event.type == KEYDOWN:
-                if event.key == K_w:
-                    player1.moveup()
-                if event.key == K_s:
-                    player1.movedown()
-                if event.key == K_a:
-                    player1.moveleft()
-                if event.key == K_d:
-                    player1.moveright()
-                
-            elif event.type == KEYUP:
-                #if event.key == K_a or event.key == K_w or event.key == K_s or event.key == K_d:
-                #    player1.movepos = [0,0]
-                #    player1.state = "still"
-
-                if event.key == K_a or event.key == K_d:
-                    player1.stopmovehoriz()
-
-                if event.key == K_w or event.key == K_s:
-                    player1.stopmovevert()
-                    
-                
-
-        
-        screen.blit(background, player1.rect, player1.rect)
-        
-        playersprites.update()
-        playersprites.draw(screen)
-        pygame.display.flip()
-
-
-if __name__ == '__main__': main()
-
-while 1:
-    for event in pygame.event.get():
-        if event.type == pygame.QUIT: sys.exit()
-
-    #screen.fill(black)
-    #screen.blit(ball, ballrect)
-
-    pygame.display.flip()
